requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
python-dotenv==1.1.0

# Testing & linting
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
ruff==0.9.3
